Cleans up continuous form biogeochemistry validation example
